# docs/swagger.yaml
basePath: /
definitions:
...
  entities.PendingWithdrawProposal:
    properties:
      aievaluation:
        type: string
      createdAt:
        type: string
      creator:
        type: string
      description:
        type: string
      id:
        type: string
      poolID:
        type: string
      poolName:
        type: string
      profileID:
        type: string
      proofBlobID:
        type: string
      purpose:
        type: string
      reviewedBy:
        type: string
      status:
        type: string
      target:
        type: string
      updatedAt:
        type: string
      withdrawAmount:
        format: int64
        type: integer
    type: object
...
  request.CreatePendingWithdrawProposalRequest:
    properties:
      description:
        type: string
      pool_id:
        type: string
      proof_blob_id:
        type: string
      withdraw_amount:
        maximum: 20000000
        minimum: 10000
        type: integer
    required:
      - description
      - pool_id
      - withdraw_amount
    type: object
...
  request.ExecuteTransactionRequest:
    properties:
      center_req:
        type: string
      proposal_id:
        type: string
      registration_req:
        type: string
      signature:
        type: string
      tx_bytes:
        type: string
      upload_child_req:
        type: string
    required:
      - signature
      - tx_bytes
    type: object
...
  response.BuildTransactionResponse:
    properties:
      center_req:
        type: string
      proposal_id:
        type: string
      registration_req:
        type: string
      tx_bytes:
        type: string
      upload_child_req:
        type: string
    type: object
...
  response.MessageAPIResponse:
    properties:
      message:
        type: string
    type: object
...
  response.PaginationDataResponse:
    properties:
      amount:
        type: integer
      data: {}
      page:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: agrotrust-server-production.onrender.com
info:
  contact: {}
  description: API for AgroTrust Server
  title: AgroTrust Server API
  version: "1.0"
paths:
...
  /pending-withdraw-proposals:
    get:
      consumes:
        - application/json
      description:
        Retrieve a list of pending withdraw proposals with optional query
        filters
      parameters:
        - in: query
          name: creator
          type: string
        - in: query
          name: keyword
          type: string
        - in: query
          name: max_amount
          type: integer
        - in: query
          name: min_amount
          type: integer
        - in: query
          name: page
          type: integer
        - in: query
          name: page_size
          type: integer
        - in: query
          name: reviewer
          type: string
        - in: query
          name: sort_criteria
          type: string
        - in: query
          name: sort_order
          type: string
        - in: query
          name: status
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.PaginationDataResponse"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: List pending withdraw proposals
      tags:
        - Pending Withdrawal
    post:
      consumes:
        - application/json
      description: Create a new pending withdraw proposal
      parameters:
        - description: "Pending Withdraw Proposal details (e.g., "
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/request.CreatePendingWithdrawProposalRequest"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/response.BuildTransactionResponse"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: Create a new pending withdraw proposal
      tags:
        - Pending Withdrawal
  /pending-withdraw-proposals/{id}:
    get:
      consumes:
        - application/json
      description:
        Retrieve details of a specific pending withdraw proposal by its
        ID
      parameters:
        - description: Pending Withdraw Proposal ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/entities.PendingWithdrawProposal"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: Get a pending withdraw proposal
      tags:
        - Pending Withdrawal
  /pending-withdraw-proposals/{id}/approve:
    post:
      consumes:
        - application/json
      description:
        Approve a pending withdraw proposal to prepare and build a transaction
        to publish this proposal on chain
      parameters:
        - description: Oending Withdraw Proposal ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.BuildTransactionResponse"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: Approve a pending withdraw proposal
      tags:
        - Pending Withdrawal
  /pending-withdraw-proposals/{id}/refuse:
    post:
      consumes:
        - application/json
      description: Refuse a pending withdraw proposal
      parameters:
        - description: Oending Withdraw Proposal ID
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/response.BuildTransactionResponse"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: Refuse a pending withdraw proposal
      tags:
        - Pending Withdrawal
...
  /tx/execute:
    post:
      consumes:
        - application/json
      description:
        Processes a transaction request and executes it on the Sui Network
        blockchain
      parameters:
        - description: Transaction execution details
          in: body
          name: request
          required: true
          schema:
            $ref: "#/definitions/request.ExecuteTransactionRequest"
      produces:
        - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "400":
          description: Invalid data. Please try again.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "401":
          description: You have no rights to access this action.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
        "500":
          description:
            There is something wrong in the system during the process.
            Please try again later.
          schema:
            $ref: "#/definitions/response.MessageAPIResponse"
      security:
        - BearerAuth: []
      summary: Execute an on-chain transaction
      tags:
        - transactions
...
schemes:
  - https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"